# internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Estoque microservice API documentation.",
        "title": "API Estoque",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/stock-move": {
            "get": {
                "description": "Retorna todas as movimentações de estoque",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-moves"
                ],
                "summary": "Listar movimentações de estoque",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova movimentação de estoque",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-moves"
                ],
                "summary": "Criar movimentação de estoque",
                "parameters": [
                    {
                        "description": "Movimentação de Estoque",
                        "name": "stockMove",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stockmoves.StockMove"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/stock-move/by-product/{idProduct}": {
            "get": {
                "description": "Retorna todas as movimentações de estoque de um produto específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-moves"
                ],
                "summary": "Listar movimentações por produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do Produto",
                        "name": "idProduct",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/stock-move/by-warehouse-product/{idWarehouse}/{idProduct}": {
            "get": {
                "description": "Retorna todas as movimentações de estoque filtradas por armazém e produto",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-moves"
                ],
                "summary": "Listar movimentações por armazém e produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do Armazém",
                        "name": "idWarehouse",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID do Produto",
                        "name": "idProduct",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/stock-move/by-warehouse/{idWarehouse}": {
            "get": {
                "description": "Retorna todas as movimentações de estoque de um armazém específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-moves"
                ],
                "summary": "Listar movimentações por armazém",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do Armazém",
                        "name": "idWarehouse",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/stock-move/{id}": {
            "get": {
                "description": "Retorna uma movimentação de estoque específica pelo seu ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-moves"
                ],
                "summary": "Buscar movimentação de estoque por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID da Movimentação",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httpresponse.Response": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "stockmoves.StockMove": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "qty_moved": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
    }
}